Make sure wasm-node is treated as an external
This is a strange one as the wasm-pack generated ./wasm-node folder
contains a file with this import:

const { TextDecoder } = require(String.raw`util`);

This apparently breaks the way how Webpack interprets this import and we
end up not finding the "util" module when using this build.

To prevent this we keep Webpack out of this matter and import
'wasm-node' unbundled / as an external module living in the library
folder.