Bring back private_key methods in KeyPair struct
I realized we don't need to change the API when changing the fields to
now wrap the whole Ed25519 keypair in our KeyPair struct.

The old API had the advantage to only work with the private key to
restore the key pair. Its only a small difference but does not need a
redundant public part, saves some bytes and is more elegant :-)