Restructure args, forge and message traits
Arguments for spaces, encryption and auth are now moved into their
regarding `message` modules.

`ForgeArgs` have been renamed to `SpacesArgs` as they are also used
outside of forging as message.

The traits for messages have been separated into a `AuthoredMessage`
which gives us the id and author of a message while the `SpacesMessage`
trait returns the arguments required to run the `p2panda-spaces`
system.

There's a minor refactor around creating the `SpacesArgs` object out of
`EncryptionArgs` and `AuthArgs`, hopefully this will make it more
readable now.