WIP: Process remote application messages
This is still not working as the encryption "process" method fails with
stating that "we're not part of the group yet". This is because the
DGM::members method currently returns an empty vec. We need to get the
correct state from auth in the DGM to correctly report the state.