Assemble encryption state, remove ephemeral sig
We need to "assemble" the encryption state (`encryption_y`) from various
parts: Key registry and the key manager are globally used across all
spaces, so we pass this in for every space individually. The dgm state
also gets populated "anew" each time, using the current auth state.

Unfortunately we also need to remove the signing of the create message
with the ephemeral key as DCGKA looks at the "sender" of this message to
determine where to get the pre-keys from.