Basic `p2panda-spaces` API with methods to create spaces and process messages (#776)
* Initial sketches

* Break up into separate files

* WIP: Sketching out new transaction-capable stores

* Note on protecting against multiple transactions being open at the same time

* Rename receive -> process & fix some generic parameters

* Use inner mutability pattern for Manager and impl Clone

* WIP: create method on Space

* WIP: Establish auth state when creating space

* WIP: Introduce types for encryption

* WIP: establish encryption group state in space.create method

* Use RefCell and Rc for Manager instead, introduce own key manager and registries

* WIP: create encryption group state and start implementing forge args

* WIP: Forge message and process it in auth group

* Add more todos and comments

* Fix recursion in Display impl

* Formatting in Cargo.toml

* Move auth and encryption-related code into related mods

* Silence clippy warnings on large types

* Get key registry and -manager state from store

* Key- and space stores

* Use async locks again to deal with ref cell across await points

* Test utilities with in-memory store

* Always derive Clone in key-manager and -registry

* Use key-manager and -registry from p2panda-encryption

* Methods to add or get key bundle

* Introduce member struct and members method on space

* Publish encrypted messages

* Assert message authors

* Restructure args, forge and message traits

Arguments for spaces, encryption and auth are now moved into their
regarding `message` modules.

`ForgeArgs` have been renamed to `SpacesArgs` as they are also used
outside of forging as message.

The traits for messages have been separated into a `AuthoredMessage`
which gives us the id and author of a message while the `SpacesMessage`
trait returns the arguments required to run the `p2panda-spaces`
system.

There's a minor refactor around creating the `SpacesArgs` object out of
`EncryptionArgs` and `AuthArgs`, hopefully this will make it more
readable now.

* Remove unused file

* WIP: Prepare process method on manager

* Handle unkown spaces outside of storage layer

Store errors should only be related to problems with the storage layer
itself.

We handle "tried to access unknown space" outside of it now and return
an `Option` instead when getting a space by id.

* Route messages to correct space or error

* Use seedable rng to create private keys

* Move tests into own module, add process test

* Manager method to get space

* Convert forged message back into encryption types

* WIP: Process remote control messages

This completes the flow to process control messages coming from other
peers. It's still a bit rough on the edges. Some things are left:

* EncryptionGroup::receive calls "members" internally to find out if
  we've been added to the group with processing this message. This
  doesn't work right now since DGM::members is a no-op
* To welcome someone into a group (via a "create") they need to
  establish the auth and encryption state first. The code for this is
  repeated right now, probably something to factor out
* The orderer is not respecting welcome state yet

* WIP: Process remote application messages

This is still not working as the encryption "process" method fails with
stating that "we're not part of the group yet". This is because the
DGM::members method currently returns an empty vec. We need to get the
correct state from auth in the DGM to correctly report the state.

* Expose fields in encryption states to handle dgm

We need to be able to assemble state objects from the outside to allow
more flexibility for different persistance approaches.

Like this we can "fill in" the latest DGM state from other sources as
soon as we're processing encryption states.

* Assemble encryption state, remove ephemeral sig

We need to "assemble" the encryption state (`encryption_y`) from various
parts: Key registry and the key manager are globally used across all
spaces, so we pass this in for every space individually. The dgm state
also gets populated "anew" each time, using the current auth state.

Unfortunately we also need to remove the signing of the create message
with the ephemeral key as DCGKA looks at the "sender" of this message to
determine where to get the pre-keys from.

* Return events with plaintext

* Forgot to add event module

* Make clippy happy

* Make all lib modules public

* Derive serde traits behind feature flag in p2panda-auth

* Derive Clone, Serialize and Deserialize where required

* Add conversion trait into PublicKey for ActorId

* Pin compatible rand version in p2panda-auth

* Remove TODO note

It's fine to clone here

* Export core types from lib-level

* Improve ActorId API

* Use From instead of Into to convert from ActorId to PublicKey
* Introduce `as_bytes` method on ActorId
* Make placeholder methods only visible to crate

* Methods to convert from bytes to ActorId

* Make clippy happy

* Conversion methods for OperationId

* Convert from any byte slice to ActorId

* Remove unnecessary mut

* Member constructor should be public

* Test to check actor id conversion methods

---------

Co-authored-by: Sam Andreae <contact@samandreae.com>