* Handle auth and encryption message dependencies
Dependencies required so that messages can be processed in causal / partial order by any peer. Both p2panda-auth and p2panda-encryption have their own ordering requirements specified through dependencies.

In order to support these requirements we introduce auth and encryption dependency fields to message types and implement the logic to attach the relevant dependencies during message construction.

Calculating what dependencies should be attached is the responsibility of an orderer (both auth and encryption require one), in this commit we do not implement the logic for keeping the orderer aware of current state, it only attaches dependencies to messages based on it's own internal state.