* Implement "add" method on space
This is the method for adding new members to a space, with an associated access level. Members with "read" access or above will be added to the encryption scope.

In this commit we only support adding individual members.

When making the encryption DGM aware of the new post-add group state, we  calculate the next membership state for a space by processing a "fake" operation before we forged the real one.