* Introduce AuthorityGraphs struct
Authority graphs are required for detecting concurrent mutual remove cycles which may include authority delegations. The authority graph is concurrency aware as it knows about the global auth dependency graph. It contains contains nodes made up of (actor, operation id) tuples,  and edges concurrent removals or non-concurrent delegations.

Once the authority graph has been populated with removals and delegations, it can be queried to understand if an operation is part of a cycle for a particular group.