* Improvements to strong remove resolver rules and detect nested group cycles
Strong remove resolver now detects and handles concurrent re-adds and mutual-remove cycles. The latter makes use of the previously added AuthorityGraphs.

In GroupCrdt nested group cycles are detected and rejected when calling process. A max depth check is introduced to when querying group state to protect against accidental cycles which may still occur as a result of concurrent operations.