Maintain dependencies for auth and encryption
Add dependency graph to both auth and orderer states so that dependencies for both can be maintained separately. So far we only track control messages, but later we will add application message dependencies.

Dependencies are updated when any local methods are called (create, add) or we receive and process a remote message.