Handle pull-only members
Space members with only Pull access should be added to the auth group, but _not_ the encryption group. A pre-processing step is introduced in this commit to handle that case.

As this change complicated the spaces logic a little more I now refactored re-usable code into individual methods for code-comprehensibility. State objects are passed into these method (rather than retrieving it from the store) when mutation occurs in order to protect against dirty DB reads and keep the (internal) API as explicit as possible.