Process encryption messages after welcome
We want to queue up all group encryption control and application messages until we have actually been welcomed into the group. This can happen in a "Create" or an "Add" message. In the case of the latter, many other control messages may have arrived before we actually join the group and should process them.

To do this we implement the set_welcome method on the encryption orderer and only return next messages when the DGM flagged us as "welcomed".

Additionally a check is added which returns an error if space.publish is called before we are welcomed to a group.