`p2panda-spaces`: introduce `space.add` method (#785)
* * Handle auth and encryption message dependencies

Dependencies required so that messages can be processed in causal / partial order by any peer. Both p2panda-auth and p2panda-encryption have their own ordering requirements specified through dependencies.

In order to support these requirements we introduce auth and encryption dependency fields to message types and implement the logic to attach the relevant dependencies during message construction.

Calculating what dependencies should be attached is the responsibility of an orderer (both auth and encryption require one), in this commit we do not implement the logic for keeping the orderer aware of current state, it only attaches dependencies to messages based on it's own internal state.

* * Implement "add" method on space

This is the method for adding new members to a space, with an associated access level. Members with "read" access or above will be added to the encryption scope.

In this commit we only support adding individual members.

When making the encryption DGM aware of the new post-add group state, we  calculate the next membership state for a space by processing a "fake" operation before we forged the real one.

* Clippy & fmt

* Clippy

* TODO comments

* fmt

* Maintain dependencies for auth and encryption

Add dependency graph to both auth and orderer states so that dependencies for both can be maintained separately. So far we only track control messages, but later we will add application message dependencies.

Dependencies are updated when any local methods are called (create, add) or we receive and process a remote message.

* DGM::from_welcome is a no-op

* Method doesn't need to be public

* Comment out test for processing remote messages

* fmt

* Clippy

* Handle pull-only members

Space members with only Pull access should be added to the auth group, but _not_ the encryption group. A pre-processing step is introduced in this commit to handle that case.

As this change complicated the spaces logic a little more I now refactored re-usable code into individual methods for code-comprehensibility. State objects are passed into these method (rather than retrieving it from the store) when mutation occurs in order to protect against dirty DB reads and keep the (internal) API as explicit as possible.

* Delete dead code in tests

* Rename persist_space method to set_state

---------

Co-authored-by: adz <x12@adz.garden>