discovery: Enable nonblocking for unbound socket (#794)
* tests: Add test to make sure an unbound socket for mdns can be created

This test is expected to fail and the next commit fixes it.

* discovery: Enable nonblocking for unbound socket

In case it's not possible to create a bound socket, a unbound socket is
created. The unbound socket also needs to enable nonblocking.
This causes a crash in reflection when the device doesn't have any
network connection.

Related: https://github.com/p2panda/reflection/issues/110