Introduce Checkpoint
A checkpoint is a state-vector describing a "cut" in an operation graph.

The latest state of a graph would be the "frontier".

We use checkpoints to describe from where we want to continue streaming
data. It can also be used to send the state vector to another peer to
understand the set difference (for example for a sync protocol).