`p2panda-auth`: handle strong removal edge cases and cycles in nested groups, improve concurrent re-adds (#788)
* * Introduce AuthorityGraphs struct

Authority graphs are required for detecting concurrent mutual remove cycles which may include authority delegations. The authority graph is concurrency aware as it knows about the global auth dependency graph. It contains contains nodes made up of (actor, operation id) tuples,  and edges concurrent removals or non-concurrent delegations.

Once the authority graph has been populated with removals and delegations, it can be queried to understand if an operation is part of a cycle for a particular group.

* Introduce new graph helper methods

* * Improvements to strong remove resolver rules and detect nested group cycles

Strong remove resolver now detects and handles concurrent re-adds and mutual-remove cycles. The latter makes use of the previously added AuthorityGraphs.

In GroupCrdt nested group cycles are detected and rejected when calling process. A max depth check is introduced to when querying group state to protect against accidental cycles which may still occur as a result of concurrent operations.

* fmt & Clippy

* fmt fuzz

* fmt

* Add CHANGELOG entry

* Move constant to top of file

* Some more doc strings

* fmt

* Improve CHANGELOG entry

* Revert change in fuzz/groups_message

* Optimise conditional statements