New stream API for layer event processors (#804)
* First experimental approach with processing layers

* Single-threaded approach

* Methods and implementations to work with Stream

* Rename chained -> composed

* A note on single-threaded design

* A note on ugly error types of the buffer layer

* Improve stream design by embedding buffer logic

There was an potential foot-gun when combining an expensive async
processor with the Stream-based API (you can see a panic there handling
that situation). This is due to a simplified polling approach where we
don't park the futures for contiuous polling. The solution is to either
take care of buffering in the processor implementation directly or wrap
it around the BufferedProcessor helper.

The solution this commit takes is to have that directly be handled in
the stream.

Another improvement is removing the `Terminated` error type in
BufferedProcessor is a bit ugly as technically it will never really
happen (when the channel is dropped everything else is gone as well).

* Minor clean up

* tokio-stream is a dev dependency

* Correct mod-level doc-string syntax

* Fix typo: seize -> cease

* Apply grammar fixes from code review

Co-authored-by: glyph <gnomad@cryptolab.net>

* One more doc-string improvement

---------

Co-authored-by: glyph <gnomad@cryptolab.net>