Adjust PartialOrder trait, remove operation impl
The trait bounds on the trait are only required by the in-memory
implementation.

The error type is added as an associated trait type as it's dependent on
the concrete storage layer implementation.

The operation implementation was removed as this will now be part of the
new orderer processor implementation in p2panda-stream-next.