`p2panda-spaces`: global shared auth state and basic groups API (#796)
* Global shared auth state and basic group API

There is one global auth state which is shared by all spaces. This allows for groups to be re-used across multiple spaces. Keeping spaces in sync with the global poses several challenges considering the auth state itself will be mutated concurrently by multiple actors.

The other property we need to uphold is per-space atomicity, meaning that any encryption message must be processed along with it's accompanying auth message, or not at all.

These two requirements can be satisfied if whenever an auth group membership change occurs, an auth message is forged and published towards a global auth state _and_ a space message (containing encryption messages) is forged and published to all known spaces referencing this message.

With this approach all spaces are effected when the global state changes, and per-space atomicity is retained as encryption state changes will always occur along with their related auth state change.

There are special concurrency cases which need to be handled (auth state changes needing to be retroactively applied to concurrently created spaces), these will be dealt with in further commits.

* Clippy & fmt

* Remove incorrect group doc string

* Make space ids generic

This commit introduces generic space ids. This allows users to bring their own space identifier type. It requires one change to the public API which is that a space id must be provided when calling create_space(..) on the manager.

* Helper methods for adding group member types to actor ids

* Helper method for initialising space state from existing auth state

* Refactoring and comment improvement

* Clippy

* Remove unnecessary clippy allows

* Doc strings for all public methods and add TODOs

* Sync auth changes to all local spaces

When we perform a change to the shared auth state we need to sync all spaces we know about. We do this by processing the resulting auth message on them as well.

* Move space membership message handling code into own method

* Auth message is not optional for membership messages

* Test for shared auth state

* fmt

* Handle manager errors occuring inside Group

* Fix: inject latest key bundles into space state

When fetching space state from the store we need to inject the latest state for both the PreKeyManager and PreKeyRegistry as changes may have occurred.

* Derive encryption message from space message

When processing a group membership change on a space locally (potentially effecting multiple members), collect all resulting direct messages and publish only these (no encryption control message) on the space message. When receiving and processing a space message, derive the encryption control message from the corresponding auth message and include all direct messages.

* Remove TODO and improve doc string

* Include application messages in encryption orderer state

* Strong format manager error to avoid error type cycles

* Doc, formatting and naming improvements

* Doc string correction

Co-authored-by: glyph <gnomad@cryptolab.net>

* Comment correction

Co-authored-by: glyph <gnomad@cryptolab.net>

* fmt

* More explicit method names

---------

Co-authored-by: glyph <gnomad@cryptolab.net>