Introduce process_batch method to Manager
Introduce a new method for processing batches of messages in one go.

Additionally, methods for detecting spaces which have become out-of-sync with the global shared auth state and "repairing" them (apply messages from the shared auth state a space is missing) are introduced and called after a batch of messages are processed.