`p2panda-spaces`: api for detetcting and publishing missing auth pointers to spaces (#824)
* Introduce process_batch method to Manager

Introduce a new method for processing batches of messages in one go.

Additionally, methods for detecting spaces which have become out-of-sync with the global shared auth state and "repairing" them (apply messages from the shared auth state a space is missing) are introduced and called after a batch of messages are processed.

* fmt & Clippy

* Don't repair in process_batch

* Fix: record processed auth messages when building space state

* Tests

* Comment clarification

* Update tests

* Improve repair_spaces doc string

* Remove process_batch method

* Test processing duplicate auth reference messages on space

* Put doc string diagrams in text blocks

* Fix after rebase

* Don't process duplicate auth pointers on spaces auth state