Fix ordering for events updating our own transport info (#904)
* Remove incorrect comment

We actually need "publish" for maintaining our own node info which is
used during discovery.

* Fix updated transport race conditions

Transport info was sometimes not updated as very fastly updated endpoint
addresses (coming from iroh) would not be caught since our timestamp
precision is not accounting for sub-second changes.

Introducing microsecond precision for timestamps would be probably
enough, we're still adding a "hybrid" solution, using lamport timestamps
as a tie-breaker in case the microsecond timestamp is the same.

This hybrid solution accounts especially for mocked system time setups
in test environments.

We're getting the address from iroh via a blocking call (it's the
"publish" method on their `Discovery` trait), so we need to spawn an
async tokio task there to get access to the address book. Since there's
multiple tasks spawned which race against each other every time iroh
changes our endpoint address, the code inside of them might be executed
in "wrong" order. I've added a semaphore to enforce strict ordering of
the events as if this would be a blocking method.

* Fix clippy warnings

* Only update our info if it really changed

* Remove accidentially removed workspace crate

* Minor formatting